feat: implement JWT authentication, user management, and church functionalities

+++ src/application/usecases/user/delete-user.usecase.ts
@@ -1,12 +1,13 @@
-import { Injectable } from '@nestjs/common';
+import { ForbiddenException, Injectable } from '@nestjs/common';
 import { UserRepository } from 'src/domain/repositories/user.repository';
 
 @Injectable()
 export class DeleteUserUsecase {
   constructor(private readonly userRepository: UserRepository) {}
 
-  async execute(id: string): Promise<void> {
-    await this.userRepository.findById(id);
+  async execute(id: string, churchId: string): Promise<void> {
+    const user = await this.userRepository.findById(id);
+    if (user.churchId !== churchId) throw new ForbiddenException();
     return this.userRepository.delete(id);
   }
 }
